# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Adpoint API",
        "contact": {
            "email": "alkhamilnaz@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        },
        {
            "url": "https://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        }
    ],
    "paths": {
...
        "/subcategories": {
            "get": {
                "tags": [
                    "Sub Categories"
                ],
                "summary": "Display a listing of the subcategories",
                "operationId": "list subcategories",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subcategory/add": {
            "post": {
                "tags": [
                    "Sub Categories"
                ],
                "summary": "Add Subcategory",
                "operationId": "Add Subcategory",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubCategorySchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subcategory/{id}": {
            "get": {
                "tags": [
                    "Sub Categories"
                ],
                "summary": "Display a listing of the subcategory by id",
                "operationId": "list subcategory by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of subcategory to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sub Categories"
                ],
                "summary": "Delete subcategory by id",
                "operationId": "Delete subcategory by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "subcategory id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subcategory/edit/{id}": {
            "put": {
                "tags": [
                    "Sub Categories"
                ],
                "summary": "Edit of the subcategory by id",
                "operationId": "Edit subcategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of edit subcategory",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubCategorySchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/upload": {
            "post": {
                "tags": [
                    "Upload Media"
                ],
                "summary": "Upload image and file",
                "operationId": "Upload image and file",
                "parameters": [
                    {
                        "name": "parent",
                        "in": "query",
                        "description": "directory parent ( users/brands/categories/products )",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "child",
                        "in": "query",
                        "description": "directory child ( users:ktp/npwp | brands:logo | categories:banner/icon | products:photos/thumbnail/featured/flashdeal/meta/pdf  )",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "file",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display a listing of the user",
                "operationId": "list user",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display a listing of the user by id",
                "operationId": "list user by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SubCategorySchema": {
                "title": "Sub Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "type": "string",
                        "example": "Horizontal"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "slug": {
                        "type": "string",
                        "example": "Videotron"
                    },
                    "brands": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "enum": [
                            "7",
                            "4",
                            "3"
                        ]
                    },
                    "meta_title": {
                        "type": "string",
                        "example": "Videotron"
                    },
                    "meta_description": {
                        "type": "string",
                        "example": "Videotron"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}